PW-2171 Generic Component Backend for SG (#258)
backend for SG
