+++ cartridges/int_adyen_controllers/cartridge/templates/default/components/order/orderdetailsemail.isml
