+++ cartridges/int_adyen_overlay/cartridge/scripts/util/AdyenHelper.ds
@@ -34,7 +34,7 @@ var __AdyenHelper: Object = {
     LOADING_CONTEXT_TEST: "https://checkoutshopper-test.adyen.com/checkoutshopper/",
     LOADING_CONTEXT_LIVE: "https://checkoutshopper-live.adyen.com/checkoutshopper/",
 
-    CHECKOUT_COMPONENT_VERSION: "3.6.1",
+    CHECKOUT_COMPONENT_VERSION: "3.7.0",
     VERSION: "19.2.2",
 
     getService: function (service: String): Object {
