--- cartridges/int_adyen_overlay/cartridge/templates/default/adyen_confirmation.isml
@@ -1,50 +1,0 @@
-<!--- TEMPLATENAME: adyen_confirmation.isml --->
-<!---
-?merchantReference=00000100&skinCode=7uIGphb3&shopperLocale=de&paymentMethod=visa&authResult=REFUSED&pspReference=8513022639156767&merchantSig=OFMuA7Ny5%2BnRpK%2BJ0JY7CFLDie8%3D
---->
-
-<isdecorate template="checkout/pt_orderconfirmation">
-<isinclude template="util/modules"/>
-<isinclude template="util/reporting/ReportOrder.isml"/>
-<iscomment>
-	This template visualizes the order confirmation page. Note, that it
-	uses a different decorator template.
-	It displays the order related information, such as the order number,
-	creation date, payment information, order totals and shipments of
-	the order.
-</iscomment>
-
-<div class="orderconfirmation">
-
-
-
-	<div class="thankyoumessage">
-
-	<isif condition = ${pdict.CurrentHttpParameterMap.authResult == "AUTHORISED"}>
-		<h2>Your payment was successful</h2>
-		<h2>${Resource.msg('confirmation.thankyou','checkout',null)}</h2>
-	<iselse>
-		<h2>Your payment has failed</h2>
-	</isif>
-<hr size='1'></hr>
-
-		
-		<p>${Resource.msg('confirmation.message','checkout',null)}</p>
-		<p>${Resource.msg('confirmation.contact','checkout',null)}</p>
-		<p><a class="printpage">${Resource.msg('confirmation.print','checkout',null)}</a></p>
-	</div><!-- END: thankyoumessage -->
-	<div class="orderconfirmationdetails">
-		<isorderdetails order="${pdict.Order}"/>
-	</div>
-	<div class="actions">
-		<a href="${URLUtils.http('Cart-ContinueShopping')}">${Resource.msg('confirmation.returnshop','checkout',null)}</a>
-	</div>
-</div><!-- END: orderconfirmation -->
-<script type="text/javascript"><!--
-jQuery(document).ready(function() {
-	jQuery("a.printpage").click(function() {
-		window.print();
-	});
-});
-//--></script>
-</isdecorate>
