--- cartridges/int_adyen_overlay/cartridge/templates/default/afterpay.isml
@@ -1,2 +1,0 @@
-<iscontent type="text/html" encoding="off" />
-<isprint value="${pdict.OpenInvoiceResponse}" />
