--- cartridges/int_adyen_overlay/cartridge/templates/default/afterpayerror.isml
@@ -1,1 +1,0 @@
-ERROR
