--- cartridges/int_adyen_overlay/cartridge/templates/default/checkoutComponent.isml
@@ -1,73 +1,0 @@
-<iscontent type="text/html" charset="UTF-8" compact="true"/>
-
-<isset name="openInvoiceMethodDetails" value="${AdyenHelper.filterOutOpenInvoiceComponentDetails(methodDetails)}" scope="page">
-<div id="component_${method.type}" class="checkoutComponent"></div>
-<script type="text/javascript">
-
-switch("${method.type}"){
-	case "ideal": 
-		var paymentMethodDetails = JSON.parse("${methodDetails}".replace(/&quot;/g, '\"'));
-		var component = document.getElementById("component_ideal");
-		var idealContainer = document.createElement("div");
-       	var idealComponent = AdyenCheckoutObject.create('ideal', {
-               details: paymentMethodDetails.details
-           });
-           component.append(idealContainer);
-           idealComponent.mount(idealContainer);
-        break;
-     case "klarna":
-     	var paymentMethodDetails = JSON.parse("${openInvoiceMethodDetails}".replace(/&quot;/g, '\"'));
-     	if(paymentMethodDetails){
-	     	var component = document.getElementById("component_klarna");
-			var klarnaContainer = document.createElement("div");
-	       	var klarnaComponent = AdyenCheckoutObject.create('klarna', {
-	               details: paymentMethodDetails,
-	               visibility: {
-	               		personalDetails: "editable"
-	               },
-	           });
-	           component.append(klarnaContainer);
-	           klarnaComponent.mount(klarnaContainer);
-	
-	         var billingCountry = document.getElementsByClassName("input-select country")[0].value;
-	         if(isNordicCountry(billingCountry.toUpperCase())){
-	             var ssnContainer = document.createElement("div");
-	             ssnContainer.setAttribute("id", "ssn_klarna");
-	             var socialSecurityNumberLabel = document.createElement("span");
-	             socialSecurityNumberLabel.textContent = "Social Security Number";
-	             var socialSecurityNumber = document.createElement("input");
-	             socialSecurityNumber.setAttribute("id", "ssnValue");
-	             socialSecurityNumber.setAttribute("class", "adyen-checkout__input");
-	             socialSecurityNumber.setAttribute("type", "text");
-	             
-	             ssnContainer.append(socialSecurityNumberLabel);
-	             ssnContainer.append(socialSecurityNumber);
-	             klarnaContainer.append(ssnContainer);
-	         }
-         }
-        break;
-     case "afterpay_default":
-     	var paymentMethodDetails = JSON.parse("${openInvoiceMethodDetails}".replace(/&quot;/g, '\"'));
-     	var component = document.getElementById("component_${method.type}");
-		var afterpayContainer = document.createElement("div");
-       	var afterpayComponent = AdyenCheckoutObject.create('afterpay', {
-               details: paymentMethodDetails,
-               visibility: {
-               		personalDetails: "editable"
-               },
-           });
-           component.append(afterpayContainer);
-           afterpayComponent.mount(afterpayContainer);
-        break;
-     default:
-		//no component
-};
-
-function isNordicCountry(country){
-    if(country === "SE"|| country === "FI" || country === "DK" || country === "NO") {
-        return true;
-    }
-    return false;
-};
-
-</script>
