--- cartridges/int_adyen_overlay/cartridge/templates/default/mail/orderrejected.isml
@@ -1,67 +1,0 @@
-<iscontent type="text/html " charset="UTF-8"/>
-<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.0 Transitional//EN">
-<iscomment>The "subject" tag contains the mail subject and can contain dynamic information, like the order number.</iscomment>
-<subject><isif condition="${!empty(pdict.MailSubject)}">${pdict.MailSubject}<iselse/><isprint value="${Resource.msg('order.orderconfirmation-email.001','order',null)}" encoding="on"/> <isprint value="${pdict.Order.orderNo}"/></isif></subject>
-<iscomment>
-	The "to" tag contains the email address of the recipient, the "from" tag the email address of the sender.
-	Each tag is to be specified max. once. Multiple email address can be separated by "," (see RFC2822).
-</iscomment>
-<to>${pdict.Order.customerEmail}</to>
-<from>${dw.system.Site.getCurrent().getCustomPreferenceValue('customerServiceEmail')}</from>
-
-<head>
-
-<isinclude template="util/modules"/>
-</head>
-<body>
-
-<table width="100%" cellpadding="0" cellspacing="0">
-	<tr>
-		<td align="center" style="${Resource.msg('email.backgroundcell','email',null)}">
-			<center>
-				<table  style="${Resource.msg('email.containertable','email',null)}">
-					<tr>
-						<td style="${Resource.msg('email.tablecell','email',null)}">
-							<a href="${URLUtils.httpHome()}" title="${Resource.msg('global.storename','locale',null)}">
-								<img src="${URLUtils.httpStatic('/images/logo.gif')}" alt="${Resource.msg('global.storename','locale',null)}" style="border:none;"/>
-							</a>
-						</td>
-						<td style="${Resource.msg('email.tablecell','email',null)}">
-							<strong>${Resource.msg('global.storename','locale',null)}</strong><br />
-							${Resource.msg('order.orderconfirmation-email.storeaddress','order',null)}<br />
-							${Resource.msg('order.orderconfirmation-email.storelocation','order',null)}<br />
-							<a href="${URLUtils.httpHome()}" title="${Resource.msg('order.orderconfirmation-email.jumptostore','order',null)}">${Resource.msg('global.storenameurl','locale',null)}</a><br />
-							${Resource.msg('order.orderconfirmation-email.storephone','order',null)}
-						</td>
-					</tr>
-					<tr>
-						<td colspan="2" style="${Resource.msg('email.tablecell','email',null)}">
-
-							<table  style="${Resource.msg('email.containertable','emailrejected',null)}">
-								<tr>
-									<th style="${Resource.msg('email.tableheader','emailrejected',null)}">${Resource.msg('email.paymentrejected','emailrejected',null)}</th>
-								</tr>
-								<tr>
-									<td style="${Resource.msg('email.tablecell','email',null)}">
-										<p>${Resource.msg('email.paymentrejectedmessage','emailrejected',null)}</p>
-										<p>${Resource.msg('confirmation.contact','checkout',null)}</p>
-									</td>
-								</tr>
-							</table>
-
-						</td>
-					</tr>
-					<tr>
-						<td style="${Resource.msg('email.tablecell','email',null)}" colspan="2">
-							<isemailorderdetails order="${pdict.Order}"/>
-						</td>
-					</tr>
-				</table>
-			</center>
-		</td>
-	</tr>
-</table>
-
-
-</body>
-</html>
