+++ cartridges/int_adyen_overlay/cartridge/templates/resources/adyen.properties
@@ -2,3 +2,4 @@ adyen.creditcard=Please enter a valid credit card information.
 myAccount.SaveCard=Please add a new payment instrument through the checkout
 terminal.selectTerminal=Please select your terminal
 terminal.noTerminals=There are no terminals connected
+adyen.paymentFailed=Payment failed, please try again.
