--- cartridges/int_adyen_overlay/cartridge/templates/resources/emailrejected.properties
@@ -1,4 +1,0 @@
-email.containertable=background:#ffffff;border:1px solid #ff0000;width:680px;
-email.tableheader=background:#ff2222;padding:5px 20px;font-size:12px;font-family:arial;text-align:left;
-email.paymentrejected=Payment was rejected
-email.paymentrejectedmessage=You payment was rejected. Contact us for more details about the failure
