--- cartridges/int_adyen_overlay/cartridge/templates/resources/hpp.properties
@@ -1,1 +1,0 @@
-hpp.empty=Click continue to review your order then continue to the hosted payment page for the transaction.
